Finish

+++ cat.itb.M6NF2Prac/Models/Client.cs
@@ -11,7 +11,7 @@ namespace cat.itb.M6NF2Prac.Models
         public virtual int Id { get; set; }
         public virtual int Code { get; set; }
         public virtual string Name { get; set; }
-        public virtual int Credit { get; set; }
+        public virtual double Credit { get; set; }
         public virtual IList<OrderProd> OrderProds { get; set; }
     }
 }
